requirements.txt
alembic==1.13.1
APScheduler==3.9.1.post1
Authlib==1.3.1
celery==5.3.6
docker_py==1.10.6
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.0
Flask_Login==0.6.3
Flask_Mail==0.9.1
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
google_api_python_client==2.132.0
librosa==0.9.2
MarkupSafe==2.1.5
noisereduce==3.0.2
numpy==1.23.4
openai_whisper==20231117
pandas==2.0.3
protobuf==5.27.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pydub==0.25.1
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
Requests==2.32.3
soundfile==0.12.1
SQLAlchemy==2.0.30
torch==2.3.0
Werkzeug==3.0.3
WTForms==3.1.2
